+++ module_tree/__init__.py
@@ -1,0 +1,1 @@
+from .tree import Tree
